Add bottom margin parameter like an existing top margin one
This commit adds "--bottom-margin" command line parameter which sets
the beginning of bottom ignored area in percent of page height by analogy
to the existing "--top-margin" parameter.
Thanks to this new parameter things like document dates or versions in the
footer can be skipped in diff output.

Use this opportunity to also write a description for both "--top-margin"
and "--bottom-margin" command line arguments.